requirements.txt
django~=3.2
numpy~=1.21.2
matplotlib~=3.3.2
pymysql~=1.0.2
pandas~=1.2.4
mplsoccer~=1.0.5
plotly~=4.14.3
pillow~=8.2.0
seaborn~=0.11.1